rockchip: edp: fix merge errors and bitrot
With this, the eDP driver now loads, talks to the VOP, and then detects
that there is no display connected. That's as far as I can test, since the
only board I own (Rock 5B) doesn't expose eDP, except maybe via a custom
passive HDMI->DP adapter cable that I don't currently own.

The purpose of this was primarily to see if it results in the same MMIO
softlocks as HDMI, since HDMI is in the same power domain as eDP.
Surprisingly it doesn't - this narrows down the HDMI issue to one of the
other clocks used by HDMI, but not eDP.