dts: rock-5b: Enable eDP for experimental purposes
The Rock 5B has no eDP output. However, eDP and HDMI share the same
physical pins on the RK3588 SoC, so in theory, a passive HDMI->(e)DP
cable could be fabricated to connect an (e)DP display.

This is strictly experimental, there are no guarantees that the Rock 5B's
circuitry won't interfere with a DP signal, since it's designed for HDMI.